backend/requirements.txt
langchain
langchain-community
langchain-openai
langchain-ollama
chromadb
python-dotenv
pypdf
fastapi
uvicorn
pandas
rank_bm25
